refactor(persistence): localize store ports and SQLite impls
