+++ src/cli/services.ts
@@ -2,7 +2,8 @@ import { AuditLogger } from "../audit/logger.js";
 import { createMetricsCollector } from "../observability/metrics.js";
 import { createObservabilityHub } from "../observability/hub.js";
 import { AlertEngine } from "../alerts/engine.js";
-import { AlertHistoryStore, type AlertHistoryStorePort } from "../alerts/history-store.js";
+import type { AlertHistoryStorePort } from "../alerts/history-store.js";
+import { AlertHistoryStore } from "../persistence/sqlite/alert-history-store.js";
 import { AutomationRunner } from "../automation/runner.js";
 import { AutomationScheduler } from "../automation/scheduler.js";
 import { TypedEventBus } from "../core/event-bus.js";
@@ -20,8 +21,10 @@ import { NotificationManager } from "../notification/manager.js";
 import { Orchestrator } from "../orchestrator/orchestrator.js";
 import { initPersistenceRuntime, type PersistenceRuntime } from "../persistence/sqlite/runtime.js";
 import { IssueConfigStore } from "../persistence/sqlite/issue-config-store.js";
-import { AutomationStore, type AutomationStorePort } from "../persistence/sqlite/automation-store.js";
-import { NotificationStore, type NotificationStorePort } from "../persistence/sqlite/notification-store.js";
+import { AutomationStore } from "../persistence/sqlite/automation-store.js";
+import type { AutomationStorePort } from "../automation/port.js";
+import { NotificationStore } from "../persistence/sqlite/notification-store.js";
+import type { NotificationStorePort } from "../notification/port.js";
 import { PathRegistry } from "../workspace/path-registry.js";
 import type { SecretsStore } from "../secrets/store.js";
 import { createTracker } from "../tracker/factory.js";
