+++ src/http/route-types.ts
@@ -8,8 +8,8 @@ import type { TypedEventBus } from "../core/event-bus.js";
 import type { RisolutoEventMap } from "../core/risoluto-events.js";
 import type { CodexControlPlane } from "../codex/control-plane.js";
 import type { RisolutoLogger } from "../core/types.js";
-import type { AutomationStorePort } from "../persistence/sqlite/automation-store.js";
-import type { NotificationStorePort } from "../persistence/sqlite/notification-store.js";
+import type { AutomationStorePort } from "../automation/port.js";
+import type { NotificationStorePort } from "../notification/port.js";
 import type { MetricsCollector } from "../observability/metrics.js";
 import type { ObservabilityHub } from "../observability/hub.js";
 import type { OrchestratorPort } from "../orchestrator/port.js";
