+++ src/persistence/sqlite/notification-store.ts
@@ -6,33 +6,11 @@ import type { NotificationDeliverySummary, NotificationRecord } from "../../core
 import { normalizeLimit } from "./query-helpers.js";
 import type { RisolutoDatabase } from "./database.js";
 import { notifications } from "./schema.js";
-
-export interface CreateNotificationInput {
-  type: string;
-  severity: NotificationRecord["severity"];
-  title: string;
-  message: string;
-  source: string | null;
-  href: string | null;
-  dedupeKey: string | null;
-  metadata: Record<string, unknown> | null;
-  createdAt: string;
-}
-
-export interface ListNotificationsOptions {
-  limit?: number;
-  unreadOnly?: boolean;
-}
-
-export interface NotificationStorePort {
-  create(input: CreateNotificationInput): Promise<NotificationRecord>;
-  list(options?: ListNotificationsOptions): Promise<NotificationRecord[]>;
-  countAll(): Promise<number>;
-  countUnread(): Promise<number>;
-  updateDeliverySummary(id: string, deliverySummary: NotificationDeliverySummary): Promise<NotificationRecord | null>;
-  markRead(id: string): Promise<NotificationRecord | null>;
-  markAllRead(): Promise<{ updatedCount: number; unreadCount: number }>;
-}
+import type {
+  CreateNotificationInput,
+  ListNotificationsOptions,
+  NotificationStorePort,
+} from "../../notification/port.js";
 
 export class NotificationStore {
   static create(db: RisolutoDatabase): NotificationStorePort {
